CLN: remove extra test files
